Rename browserHistory to history
Following conventions seen online
